SNOW-796236 Implement session.table for local testing (#883)

+++ tests/integ/scala/test_dataframe_suite.py
@@ -68,15 +68,25 @@ from tests.utils import (
 SAMPLING_DEVIATION = 0.4
 
 
-def test_null_data_in_tables(session):
+@pytest.mark.localtest
+def test_null_data_in_tables(session, local_testing_mode):
     table_name = Utils.random_name_for_temp_object(TempObjectType.TABLE)
     try:
-        Utils.create_table(session, table_name, "num int")
-        session.sql(f"insert into {table_name} values(null),(null),(null)").collect()
+        if not local_testing_mode:
+            Utils.create_table(session, table_name, "num int")
+            session.sql(
+                f"insert into {table_name} values(null),(null),(null)"
+            ).collect()
+        else:
+            session.create_dataframe(
+                [[None], [None], [None]],
+                schema=StructType([StructField("num", IntegerType())]),
+            ).write.save_as_table(table_name)
         res = session.table(table_name).collect()
         assert res == [Row(None), Row(None), Row(None)]
     finally:
-        Utils.drop_table(session, table_name)
+        if not local_testing_mode:
+            Utils.drop_table(session, table_name)
 
 
 def test_null_data_in_local_relation_with_filters(session):
@@ -136,14 +146,16 @@ def test_bulk_insert_from_collected_result(session):
         Utils.drop_table(session, table_name_copied)
 
 
-def test_write_null_data_to_table(session):
+@pytest.mark.localtest
+def test_write_null_data_to_table(session, local_testing_mode):
     table_name = Utils.random_name_for_temp_object(TempObjectType.TABLE)
     df = session.create_dataframe([(1, None), (2, None), (3, None)]).to_df("a", "b")
     try:
         df.write.save_as_table(table_name)
         Utils.check_answer(session.table(table_name), df, True)
     finally:
-        Utils.drop_table(session, table_name)
+        if not local_testing_mode:
+            Utils.drop_table(session, table_name)
 
 
 def test_create_or_replace_view_with_null_data(session):
@@ -1902,7 +1914,8 @@ def test_time_date_and_timestamp_test(session):
     assert str(session.sql("select '1970-1-1' :: Date").collect()[0][0]) == "1970-01-01"
 
 
-def test_quoted_column_names(session):
+@pytest.mark.localtest
+def test_quoted_column_names(session, local_testing_mode):
     normalName = "NORMAL_NAME"
     lowerCaseName = '"lower_case"'
     quoteStart = '"""quote_start"'
@@ -1912,13 +1925,26 @@ def test_quoted_column_names(session):
 
     table_name = Utils.random_name_for_temp_object(TempObjectType.TABLE)
     try:
-        Utils.create_table(
-            session,
-            table_name,
-            f"{normalName} int, {lowerCaseName} int, {quoteStart} int,"
-            f"{quoteEnd} int, {quoteMiddle} int, {quoteAllCases} int",
-        )
-        session.sql(f"insert into {table_name} values(1, 2, 3, 4, 5, 6)").collect()
+        if not local_testing_mode:
+            Utils.create_table(
+                session,
+                table_name,
+                f"{normalName} int, {lowerCaseName} int, {quoteStart} int,"
+                f"{quoteEnd} int, {quoteMiddle} int, {quoteAllCases} int",
+            )
+            session.sql(f"insert into {table_name} values(1, 2, 3, 4, 5, 6)").collect()
+        else:
+            session.create_dataframe(
+                [[1, 2, 3, 4, 5, 6]],
+                schema=[
+                    normalName,
+                    lowerCaseName,
+                    quoteStart,
+                    quoteEnd,
+                    quoteMiddle,
+                    quoteAllCases,
+                ],
+            ).write.save_as_table(table_name)
 
         # test select()
         df1 = session.table(table_name).select(
@@ -1991,10 +2017,12 @@ def test_quoted_column_names(session):
         assert df4.collect() == [Row(1)]
 
     finally:
-        Utils.drop_table(session, table_name)
+        if not local_testing_mode:
+            Utils.drop_table(session, table_name)
 
 
-def test_column_names_without_surrounding_quote(session):
+@pytest.mark.localtest
+def test_column_names_without_surrounding_quote(session, local_testing_mode):
     normalName = "NORMAL_NAME"
     lowerCaseName = '"lower_case"'
     quoteStart = '"""quote_start"'
@@ -2004,13 +2032,26 @@ def test_column_names_without_surrounding_quote(session):
 
     table_name = Utils.random_name_for_temp_object(TempObjectType.TABLE)
     try:
-        Utils.create_table(
-            session,
-            table_name,
-            f"{normalName} int, {lowerCaseName} int, {quoteStart} int,"
-            f"{quoteEnd} int, {quoteMiddle} int, {quoteAllCases} int",
-        )
-        session.sql(f"insert into {table_name} values(1, 2, 3, 4, 5, 6)").collect()
+        if not local_testing_mode:
+            Utils.create_table(
+                session,
+                table_name,
+                f"{normalName} int, {lowerCaseName} int, {quoteStart} int,"
+                f"{quoteEnd} int, {quoteMiddle} int, {quoteAllCases} int",
+            )
+            session.sql(f"insert into {table_name} values(1, 2, 3, 4, 5, 6)").collect()
+        else:
+            session.create_dataframe(
+                [[1, 2, 3, 4, 5, 6]],
+                schema=[
+                    normalName,
+                    lowerCaseName,
+                    quoteStart,
+                    quoteEnd,
+                    quoteMiddle,
+                    quoteAllCases,
+                ],
+            ).write.save_as_table(table_name)
 
         quoteStart2 = '"quote_start'
         quoteEnd2 = 'quote_end"'
@@ -2028,7 +2069,8 @@ def test_column_names_without_surrounding_quote(session):
         assert df1.collect() == [Row(3, 4, 5)]
 
     finally:
-        Utils.drop_table(session, table_name)
+        if not local_testing_mode:
+            Utils.drop_table(session, table_name)
 
 
 def test_negative_test_for_user_input_invalid_quoted_name(session):
@@ -2038,12 +2080,18 @@ def test_negative_test_for_user_input_invalid_quoted_name(session):
     assert "Invalid identifier" in str(ex_info)
 
 
-def test_clone_with_union_dataframe(session):
+@pytest.mark.localtest
+def test_clone_with_union_dataframe(session, local_testing_mode):
     table_name = Utils.random_name_for_temp_object(TempObjectType.TABLE)
     try:
-        Utils.create_table(session, table_name, "c1 int, c2 int")
+        if not local_testing_mode:
+            Utils.create_table(session, table_name, "c1 int, c2 int")
+            session.sql(f"insert into {table_name} values(1, 1),(2, 2)").collect()
+        else:
+            session.create_dataframe(
+                [[1, 1], [2, 2]], schema=["c1", "c2"]
+            ).write.save_as_table(table_name)
 
-        session.sql(f"insert into {table_name} values(1, 1),(2, 2)").collect()
         df = session.table(table_name)
 
         union_df = df.union(df)
@@ -2052,15 +2100,22 @@ def test_clone_with_union_dataframe(session):
         res.sort(key=lambda x: x[0])
         assert res == [Row(1, 1), Row(2, 2)]
     finally:
-        Utils.drop_table(session, table_name)
+        if not local_testing_mode:
+            Utils.drop_table(session, table_name)
 
 
-def test_clone_with_unionall_dataframe(session):
+@pytest.mark.localtest
+def test_clone_with_unionall_dataframe(session, local_testing_mode):
     table_name = Utils.random_name_for_temp_object(TempObjectType.TABLE)
     try:
-        Utils.create_table(session, table_name, "c1 int, c2 int")
+        if not local_testing_mode:
+            Utils.create_table(session, table_name, "c1 int, c2 int")
+            session.sql(f"insert into {table_name} values(1, 1),(2, 2)").collect()
+        else:
+            session.create_dataframe(
+                [[1, 1], [2, 2]], schema=["c1", "c2"]
+            ).write.save_as_table(table_name)
 
-        session.sql(f"insert into {table_name} values(1, 1),(2, 2)").collect()
         df = session.table(table_name)
 
         union_df = df.union_all(df)
@@ -2069,7 +2124,8 @@ def test_clone_with_unionall_dataframe(session):
         res.sort(key=lambda x: x[0])
         assert res == [Row(1, 1), Row(1, 1), Row(2, 2), Row(2, 2)]
     finally:
-        Utils.drop_table(session, table_name)
+        if not local_testing_mode:
+            Utils.drop_table(session, table_name)
 
 
 @pytest.mark.localtest
